Group meeting refactoring: fixing PropSpace for data parsing set up, etc

See merge request compsci307_2019spring/monopoly_team04!39

+++ src/Model/ClassicProperty.java
@@ -2,12 +2,12 @@ package Model;
 
 public class ClassicProperty implements Property {
 
-    float price;
+    double price;
     String name;
     String color;
 
 
-    public ClassicProperty(String myName, float myPrice, String myColor){
+    public ClassicProperty(String myName, double myPrice, String myColor){
         name = myName;
         price = myPrice;
         color = myColor;
@@ -17,16 +17,16 @@ public class ClassicProperty implements Property {
      * A getter method that returns the cost to place a house on this property
      * @return the cost to place a house on this property
      */
-    public float getCostofHouse(){
-        return (float)0.0;
+    public double getCostofHouse(){
+        return (double) 0.0;
     }
 
     /***
      * A getter method that returns the cost to place a hotel on this property
      * @return the cost to place a hotel on this property
      */
-    public float getCostofHotel(){
-        return (float)0.0;
+    public double getCostofHotel(){
+        return (double)0.0;
     }
 
     /***
@@ -41,7 +41,7 @@ public class ClassicProperty implements Property {
      * A getter method that returns the price of this property
      * @return the cost to purchase this property
      */
-    public float getPrice(){
+    public double getPrice(){
         return price;
     }
 
@@ -50,7 +50,7 @@ public class ClassicProperty implements Property {
      * much it costs when someone lands on this property
      * @return the total rent value to be paid
      */
-    public float calculateRent(){
-        return (float)0.0;
+    public double calculateRent(){
+        return (double) 0.0;
     }
 }
